+++ backend/api/jest.config.js
@@ -34,5 +34,13 @@ module.exports = {
     modulePathIgnorePatterns: [
         '<rootDir>/../../shared/dist',
         '<rootDir>/dist'
-    ]
+    ],
+    coverageThreshold: {
+        global: {
+            branches: 40,
+            functions: 50,
+            lines: 50,
+            statements: 50
+        }
+    }
 };